[clang] Bump TokenKey to 64 bits
This commit bumps TokenKey to 64 bits because all of its 32 bits are
in use and there is an incoming commit
(97015ad916c48ddf0fa34a4801e2851230cf5070) that adds one more
TokenKey.

rdar://175902198

+++ clang/include/clang/Basic/IdentifierTable.h
@@ -47,7 +47,7 @@ class MultiKeywordSelector;
 class SourceLocation;
 
 /// Constants for TokenKinds.def
-enum TokenKey : unsigned {
+enum TokenKey : uint64_t {
   KEYC99 = 0x1,
   KEYCXX = 0x2,
   KEYCXX11 = 0x4,
@@ -100,7 +100,7 @@ enum KeywordStatus {
 
 /// Translates flags as specified in TokenKinds.def into keyword status
 /// in the given language standard.
-KeywordStatus getKeywordStatus(const LangOptions &LangOpts, unsigned Flags);
+KeywordStatus getKeywordStatus(const LangOptions &LangOpts, uint64_t Flags);
 
 enum class ReservedIdentifierStatus {
   NotReserved = 0,
